+++ example/websocket/server/chat-multi/http_session.hpp
@@ -13,6 +13,7 @@
 #include "net.hpp"
 #include "beast.hpp"
 #include "shared_state.hpp"
+#include <boost/optional.hpp>
 #include <boost/smart_ptr.hpp>
 #include <cstdlib>
 #include <memory>
@@ -24,24 +25,15 @@ class http_session : public boost::enable_shared_from_this<http_session>
     beast::tcp_stream stream_;
     beast::flat_buffer buffer_;
     boost::shared_ptr<shared_state> state_;
-    http::request<http::string_body> req_;
 
-    struct send_lambda
-    {
-        http_session& self_;
+    // The parser is stored in an optional container so we can
+    // construct it from scratch it at the beginning of each new message.
+    boost::optional<http::request_parser<http::string_body>> parser_;
 
-        explicit
-        send_lambda(http_session& self)
-            : self_(self)
-        {
-        }
-
-        template<bool isRequest, class Body, class Fields>
-        void
-        operator()(http::message<isRequest, Body, Fields>&& msg) const;
-    };
+    struct send_lambda;
 
     void fail(beast::error_code ec, char const* what);
+    void do_read();
     void on_read(beast::error_code ec, std::size_t);
     void on_write(beast::error_code ec, std::size_t, bool close);